Move components to a separate directory
