+++ frontend/src/metabase/metadata/pages/DataModel/components/TablePicker/components/BulkTableVisibilityToggle.tsx
