+++ frontend/src/metabase/metadata/pages/DataModel/components/TablePicker/components/Results.module.css
