+++ frontend/src/metabase/metadata/pages/DataModel/components/TablePicker/components/TableVisibilityToggle.tsx
