+++ frontend/src/metabase/metadata/pages/DataModel/components/TablePicker/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./TablePicker";
