+++ frontend/src/metabase/metadata/pages/DataModel/components/TablePicker/index.ts
@@ -1,2 +1,2 @@
-export * from "./TablePicker";
+export { TablePicker } from "./components";
 export * from "./wrappers";
